Solution: 2632 in TypeScript.
Day 10 of the 30 Days of LC JavaScript Challenge.

LeetCode cannot handle this solution. I suspect that its TypeScript
installation (TypeScript 4.5.4 with Node.js 16.13.2) is too outdated
(the current versions are 5.0.4 for TypeScript and 19.9.0 for Node.js.)
It works on my local machine in Deno 1.33.0 (which uses V8 11.4.183.1
and TypeScript 5.0.3), and neither the TypeScript compiler nor the Deno
language server) have any complains.

Removing all type information results in a JavaScript version that runs
without problem in LeetCode.